+++ src/components/Lk/Settings/SettingsFields.vue
@@ -1,0 +1,18 @@
+<template>
+  <div class="settings-fields vue-component"></div>
+</template>
+
+<script>
+export default {
+  name: "settings-fields",
+  components: {},
+  data: function() {
+    return {};
+  },
+  props: [],
+  computed: {},
+  methods: {}
+};
+</script>
+
+<style scoped='scoped' lang='sass'></style>
